DELIA-51172 : During Olympic FER rewind, UVE player sends position as 0 in progress event and also rewind is not smooth.
Reason for change: For the HTTP 200 response when doing range request, the fragment stripping is impelemnted with 'repairIframes' as config param to set this. Also changed the range type from int to size_t

Test Procedure: Will update in the ticket
Risks: low

DEPS_BRANCH=rdk/components/generic/gst-plugins-rdk-aamp:4.8_p1v
DEPS_BRANCH=rdk/components/generic/aampabr:4.8_p1v
DEPS_BRANCH=rdk/components/generic/aampmetrics:4.8_p1v

+++ fragmentcollector_hls.h
@@ -323,9 +323,10 @@ public:
 	long long indexFirstMediaSequenceNumber; /**< first media sequence number from indexed manifest */
 
 	char *fragmentURI; /**< pointer (into playlist) to URI of current fragment-of-interest */
-	long long lastPlaylistDownloadTimeMS; /**< UTC time at which playlist was downloaded */
-	int byteRangeLength; /**< state for \#EXT-X-BYTERANGE fragments */
-	int byteRangeOffset; /**< state for \#EXT-X-BYTERANGE fragments */
+	long long lastPlaylistDownloadTimeMS; /**< UTC time at which playlist was downloaded */  
+        
+	size_t byteRangeLength; /**< state for \#EXT-X-BYTERANGE fragments */
+	size_t byteRangeOffset; /**< state for \#EXT-X-BYTERANGE fragments */
 
 	long long nextMediaSequenceNumber; /**< media sequence number following current fragment-of-interest */
 	double playlistPosition; /**< playlist-relative time of most recent fragment-of-interest; -1 if undefined */
